+++ mealkitary-infrastructure/adapter-paymentgateway-tosspayments/src/main/kotlin/com/mealkitary/paymentgateway/TossPaymentGatewayService.kt
@@ -1,5 +1,7 @@
 package com.mealkitary.paymentgateway
 
+import com.mealkitary.paymentgateway.payload.TossPayment
+import com.mealkitary.paymentgateway.payload.TossPaymentCancelPayload
 import com.mealkitary.reservation.domain.payment.Payment
 import com.mealkitary.reservation.domain.payment.PaymentGatewayService
 import org.springframework.stereotype.Component
