+++ mealkitary-infrastructure/adapter-paymentgateway-tosspayments/src/main/kotlin/com/mealkitary/paymentgateway/TossPaymentWebClient.kt
@@ -1,6 +1,9 @@
 package com.mealkitary.paymentgateway
 
 import com.mealkitary.common.exception.EntityNotFoundException
+import com.mealkitary.paymentgateway.payload.TossErrorResponse
+import com.mealkitary.paymentgateway.payload.TossPayment
+import com.mealkitary.paymentgateway.payload.TossPaymentCancelPayload
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpStatus
