+++ mealkitary-infrastructure/adapter-paymentgateway-tosspayments/src/main/kotlin/com/mealkitary/paymentgateway/payload/TossErrorResponse.kt
@@ -1,3 +1,3 @@
-package com.mealkitary.paymentgateway
+package com.mealkitary.paymentgateway.payload
 
 data class TossErrorResponse(val code: String, val message: String)
