+++ mealkitary-infrastructure/adapter-paymentgateway-tosspayments/src/test/kotlin/com/mealkitary/paymentgateway/TossPaymentWebClientConfirmTest.kt
@@ -3,6 +3,8 @@ package com.mealkitary.paymentgateway
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.mealkitary.common.exception.EntityNotFoundException
 import com.mealkitary.paymentgateway.codec.UrlSafeBase64Codec
+import com.mealkitary.paymentgateway.payload.TossErrorResponse
+import com.mealkitary.paymentgateway.payload.TossPayment
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.matchers.shouldBe
